deleted validations.rb rubocop fix for main.rb logic.rb
