--- bin/scraper.csv
@@ -1,22 +1,0 @@
-title,city,country,company,url
-Product Designer,Helsinki,Finland,Smartly.io,https://relocate.me//finland/helsinki/smartly-io/product-designer-5730
-Team Lead PHP,Rotterdam,The Netherlands,Coolblue,https://relocate.me//the-netherlands/rotterdam/coolblue/team-lead-php-6607
-Senior Google Cloud Engineer,Rotterdam,The Netherlands,Coolblue,https://relocate.me//the-netherlands/rotterdam/coolblue/senior-google-cloud-engineer-6608
-Senior C# Developer,Rotterdam,The Netherlands,Coolblue,https://relocate.me//the-netherlands/rotterdam/coolblue/senior-c-developer-6605
-Team Lead C#,Rotterdam,The Netherlands,Coolblue,https://relocate.me//the-netherlands/rotterdam/coolblue/team-lead-c--6606
-QA Engineer,Copenhagen,Denmark,Tactile Games,https://relocate.me//denmark/copenhagen/tactile-games/qa-engineer-6234
-Backend RoR Developer,Amsterdam,The Netherlands,Studytube,https://relocate.me//the-netherlands/amsterdam/studytube/backend-ror-developer-932
-Software Engineer,Tokyo,Japan,Mercari,https://relocate.me//japan/tokyo/mercari/software-engineer-6587
-React Native Developer,Amsterdam,The Netherlands,APC,https://relocate.me//the-netherlands/amsterdam/apc/react-native-developer-6578
-Backend Software Engineer,Amsterdam,The Netherlands,APC,https://relocate.me//the-netherlands/amsterdam/apc/backend-software-engineer-6577
-Full Stack Software Engineer - B2B services & IoT,Munich,Germany,tado°,https://relocate.me//germany/munich/tado-/full-stack-software-engineer-b2b-services-iot-5828
-"Product Manager, Technical",Tokyo,Japan,Mercari,https://relocate.me//japan/tokyo/mercari/product-manager-technical-6588
-Engineering Manager,Tokyo,Japan,Mercari,https://relocate.me//japan/tokyo/mercari/engineering-manager-6586
-Senior Quality Assurance Engineer,London,The United Kingdom,GoodNotes,https://relocate.me//the-united-kingdom/london/goodnotes/senior-quality-assurance-engineer-5753
-.NET Developer,Málaga,Spain,EPAM,https://relocate.me//spain/m-laga/epam/-net-developer-6465
-iOS Engineer,Milan,Italy,Bending Spoons,https://relocate.me//italy/milan/bending-spoons/ios-engineer-5791
-Senior Linux Administrator,Limassol,Cyprus,XM,https://relocate.me//cyprus/limassol/xm/senior-linux-administrator-5823
-Linux System Engineer,Amsterdam,The Netherlands,DEGIRO,https://relocate.me//the-netherlands/amsterdam/degiro/linux-system-engineer-5265
-Senior Web Backend Java Developer,Amsterdam,The Netherlands,DEGIRO,https://relocate.me//the-netherlands/amsterdam/degiro/senior-web-backend-java-developer-4580
-JavaScript Front End Developer,Heidelberg,Germany,Getsafe,https://relocate.me//germany/heidelberg/getsafe/javascript-front-end-developer-5440
-Ruby Backend Developer,Heidelberg,Germany,Getsafe,https://relocate.me//germany/heidelberg/getsafe/ruby-backend-developer-5124
